Move provider helper functions into framework_provider.go
Consolidates decodeStringSetOrNil and possibleValues from the separate
framework_helpers.go file directly into framework_provider.go, which is
the natural home for package-wide helpers. Deletes framework_helpers.go.
